Add check for multiple artifacts
Added check for multiple artifacts returned as the 'results' array by
the AQL query. If there are more artifacts than one, the AQL query is
wrong or something else went wrong. Log it as an error, and for now
send ArtP events for all of the returned artifacts. Will be removed
in the future, but for now we don't want to miss events.

Change-Id: If60fc97ed24592dba86cad03b76e3d1996f5927c