Add Ansible playbook for remote host
Added an Ansible playbook for the remote host, and moved all the variables
to group_vars, which in this commit are represented by .example files.
Added an Ansible template for docker-compose.yaml, which is to be used
by the playbook for a remote host.
The inventory to be used is 'hosts' ('hosts.example' in this commit)
found in dir 'ansible' instead of /etc/ansible/hosts.